MICROSOFT-ENTRA-AGENT-ID/scripts/requirements.txt
azure-ai-projects>=1.0.0
azure-identity>=1.15.0
azure-mgmt-resourcegraph>=8.0.0
python-dotenv>=1.0.0
python-dateutil>=2.8.0
requests>=2.31.0
